+++ NEWS.md
@@ -1,8 +1,8 @@
 # rcaaqs 0.3.1.9000
 
 * Finalized methods for SO2 and NO2
-* Output CAAQ status and managment status for each metric
-* Created function wrappers for each CAAQ metric
+* Output CAAQS status and managment status for each metric
+* Created function wrappers for each CAAQS metric
 * New function, `assign_airzone` to determine where sites are located in airzone shapefiles
 * Removed airzone map, use maps from `bcmaps` package instead
 * Updates to documentation